+++ Airplanes/Models/Airplane.cs
@@ -6,7 +6,7 @@ namespace Airplanes.Models
         [Key] public Guid Pid { get; set; }
         [Required] public String Pname { get; set; }
         [Required] public int Pseats { get; set; }
-        [Required] public int Pmaxspeed { get; set; }
+        [Required] public float Pmaxspeed { get; set; }
         public float Pheavyload { get; set; }
     }
 }
